Import ngtcp2_ksl changes

+++ lib/nghttp3_ksl.h
@@ -85,9 +85,9 @@ struct nghttp3_ksl_blk {
   /* prev points to the previous block if leaf field is nonzero. */
   nghttp3_ksl_blk *prev;
   /* n is the number of nodes this object contains in nodes. */
-  size_t n;
+  uint32_t n;
   /* leaf is nonzero if this block contains leaf nodes. */
-  int leaf;
+  uint32_t leaf;
   union {
     uint64_t align;
     /* nodes is a buffer to contain NGHTTP3_KSL_MAX_NBLK
